requirements.txt
accelerate
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.20
click==8.1.7
coloredlogs==15.0.1
colorama
contourpy==1.2.0
cv
cycler==0.12.1
dataclasses-json==0.6.3
datasets==2.15.0
Deprecated==1.2.14
dill==0.3.7
distro==1.8.0
document==1.0
einops
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.105.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.0
fsspec==2023.10.0
google-auth==2.25.2
googleapis-common-protos==1.62.0
greenlet==3.0.2
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.0.349
langchain-community==0.0.1
langchain-core==0.0.13
langsmith==0.0.69
lxml==4.9.4
marshmallow==3.20.1
matplotlib==3.8.2
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.3.7
openpyxl==3.1.2
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
Pillow==10.2.0
posthog==3.1.0
protobuf==4.25.1
pulsar-client==3.3.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
PyPika==0.48.9
pysqlite3==0.5.2
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-pptx==0.6.23
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy
seaborn==0.13.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
starlette==0.27.0
supervision
sympy==1.12
tensorboard
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.15.0
torch
torchvision
tqdm==4.66.1
transformers==4.39.1
transformers_stream_generator==0.0.4
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
XlsxWriter==3.1.9
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0